Add support for uploading an image ZIP and extracting them into S3

+++ api/posts.go
@@ -5,8 +5,10 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
+	"io"
 	"log"
 
+	"gocloud.dev/blob"
 	"gocloud.dev/pubsub"
 
 	"github.com/ourrootsorg/cms-server/model"
@@ -44,17 +46,32 @@ func (api API) AddPost(ctx context.Context, in model.PostIn) (*model.Post, error
 		log.Printf("[ERROR] Invalid post %v", err)
 		return nil, NewError(err)
 	}
-	// prepare to send a message
-	topic, err := api.OpenTopic(ctx, "recordswriter")
+	log.Printf("[DEBUG] Starting post %#v", in)
+	log.Printf("[DEBUG] Open recordswriter topic")
+	// prepare to send a RecordsWriter message
+	recordsWriterTopic, err := api.OpenTopic(ctx, "recordswriter")
 	if err != nil {
 		log.Printf("[ERROR] Can't open recordswriter topic %v", err)
 		return nil, NewError(err)
 	}
-	defer topic.Shutdown(ctx)
+	defer recordsWriterTopic.Shutdown(ctx)
+	log.Printf("[DEBUG] Open imageswriter topic")
+	// prepare to send a ImagesWriter message
+	imagesWriterTopic, err := api.OpenTopic(ctx, "imageswriter")
+	if err != nil {
+		log.Printf("[ERROR] Can't open imageswriter topic %v", err)
+		return nil, NewError(err)
+	}
+	defer imagesWriterTopic.Shutdown(ctx)
 	// set records status
 	in.RecordsStatus = model.PostDraft
 	if in.RecordsKey != "" {
 		in.RecordsStatus = model.PostLoading
+	}
+	// set images status
+	in.ImagesStatus = model.PostDraft
+	if in.ImagesKey != "" {
+		in.ImagesStatus = model.PostLoading
 	}
 	// insert
 	post, e := api.postPersister.InsertPost(ctx, in)
@@ -72,13 +89,33 @@ func (api API) AddPost(ctx context.Context, in model.PostIn) (*model.Post, error
 			log.Printf("[ERROR] Can't marshal message %v", err)
 			return nil, NewError(err)
 		}
-		err = topic.Send(ctx, &pubsub.Message{Body: body})
+		err = recordsWriterTopic.Send(ctx, &pubsub.Message{Body: body})
 		if err != nil { // this had best never happen
 			log.Printf("[ERROR] Can't send message %v", err)
 			// undo the insert
 			_ = api.postPersister.DeletePost(ctx, post.ID)
 			return nil, NewError(err)
 		}
+	}
+
+	if in.ImagesKey != "" {
+		// send a message to write the images
+		msg := model.ImagesWriterMsg{
+			PostID: post.ID,
+		}
+		body, err := json.Marshal(msg)
+		if err != nil { // this had best never happen
+			log.Printf("[ERROR] Can't marshal message %v", err)
+			return nil, NewError(err)
+		}
+		err = imagesWriterTopic.Send(ctx, &pubsub.Message{Body: body})
+		if err != nil { // this had best never happen
+			log.Printf("[ERROR] Can't send message %v", err)
+			// undo the insert
+			_ = api.postPersister.DeletePost(ctx, post.ID)
+			return nil, NewError(err)
+		}
+		log.Printf("[DEBUG] Sent imageswriter message '%s'", string(body))
 	}
 
 	return post, nil
@@ -97,21 +134,21 @@ func (api API) UpdatePost(ctx context.Context, id uint32, in model.Post) (*model
 		return nil, errs
 	}
 
-	var topic *pubsub.Topic
+	var recordsWriterTopic, imagesWriterTopic *pubsub.Topic
 	var msg []byte
 
 	if currPost.RecordsKey != in.RecordsKey {
 		// handle records key change
 		if currPost.RecordsStatus != in.RecordsStatus || in.RecordsStatus != model.PostDraft {
-			return nil, NewError(errors.New(fmt.Sprintf("cannot change recordsKey unless recordsStatus is Draft; status is %s", currPost.RecordsStatus)))
+			return nil, NewError(fmt.Errorf("cannot change recordsKey unless recordsStatus is Draft; status is %s", currPost.RecordsStatus))
 		}
 		// prepare to send a message
-		topic, err = api.OpenTopic(ctx, "recordswriter")
+		recordsWriterTopic, err = api.OpenTopic(ctx, "recordswriter")
 		if err != nil {
 			log.Printf("[ERROR] Can't open recordswriter topic %v", err)
 			return nil, NewError(err)
 		}
-		defer topic.Shutdown(ctx)
+		defer recordsWriterTopic.Shutdown(ctx)
 
 		in.RecordsStatus = model.PostLoading
 		msg, err = json.Marshal(model.RecordsWriterMsg{
@@ -127,12 +164,12 @@ func (api API) UpdatePost(ctx context.Context, id uint32, in model.Post) (*model
 		case (currPost.RecordsStatus == model.PostDraft && in.RecordsStatus == model.PostPublished) ||
 			(currPost.RecordsStatus == model.PostPublished && in.RecordsStatus == model.PostDraft):
 			// prepare to send a message
-			topic, err = api.OpenTopic(ctx, "publisher")
+			recordsWriterTopic, err = api.OpenTopic(ctx, "publisher")
 			if err != nil {
 				log.Printf("[ERROR] Can't open publisher topic %v", err)
 				return nil, NewError(err)
 			}
-			defer topic.Shutdown(ctx)
+			defer recordsWriterTopic.Shutdown(ctx)
 
 			var action string
 			if in.RecordsStatus == model.PostPublished {
@@ -163,14 +200,84 @@ func (api API) UpdatePost(ctx context.Context, id uint32, in model.Post) (*model
 		}
 	}
 
+	if currPost.ImagesKey != in.ImagesKey {
+		// handle images key change
+		if currPost.ImagesStatus != in.ImagesStatus || in.ImagesStatus != model.PostDraft {
+			return nil, NewError(fmt.Errorf("cannot change imagesKey unless imagesStatus is Draft; status is %s", currPost.ImagesStatus))
+		}
+		if currPost.ImagesKey != "" {
+			errs = api.DeleteImagesForPost(ctx, id)
+			if errs != nil {
+				return nil, NewError(fmt.Errorf("Failure deleting old images for post %d: %v", id, err))
+			}
+
+		}
+		// prepare to send a message
+		imagesWriterTopic, err = api.OpenTopic(ctx, "imageswriter")
+		if err != nil {
+			log.Printf("[ERROR] Can't open imageswriter topic %v", err)
+			return nil, NewError(err)
+		}
+		defer imagesWriterTopic.Shutdown(ctx)
+
+		in.ImagesStatus = model.PostLoading
+		msg, err = json.Marshal(model.ImagesWriterMsg{
+			PostID: id,
+		})
+		if err != nil {
+			log.Printf("[ERROR] Can't marshal message %v", err)
+			return nil, NewError(err)
+		}
+	} else if currPost.ImagesStatus != in.ImagesStatus {
+		// handle images status change
+		switch {
+		case (currPost.ImagesStatus == model.PostDraft && in.ImagesStatus == model.PostPublished) ||
+			(currPost.ImagesStatus == model.PostPublished && in.ImagesStatus == model.PostDraft):
+			// prepare to send a message
+			imagesWriterTopic, err = api.OpenTopic(ctx, "publisher")
+			if err != nil {
+				log.Printf("[ERROR] Can't open publisher topic %v", err)
+				return nil, NewError(err)
+			}
+			defer imagesWriterTopic.Shutdown(ctx)
+
+			var action string
+			if in.ImagesStatus == model.PostPublished {
+				in.ImagesStatus = model.PostPublishing
+				action = model.PublisherActionIndex
+			} else {
+				in.ImagesStatus = model.PostUnpublishing
+				action = model.PublisherActionUnindex
+			}
+			msg, err = json.Marshal(model.PublisherMsg{
+				Action: action,
+				PostID: id,
+			})
+			if err != nil {
+				log.Printf("[ERROR] Can't marshal message %v", err)
+				return nil, NewError(err)
+			}
+		case currPost.ImagesStatus == model.PostPublishing && in.ImagesStatus == model.PostPublishComplete:
+			in.ImagesStatus = model.PostPublished
+		case currPost.ImagesStatus == model.PostUnpublishing && in.ImagesStatus == model.PostUnpublishComplete:
+			in.ImagesStatus = model.PostDraft
+		case currPost.ImagesStatus == model.PostLoading && in.ImagesStatus == model.PostLoadComplete:
+			in.ImagesStatus = model.PostDraft
+		default:
+			msg := fmt.Sprintf("[ERROR] cannot change images status from %s to %s", currPost.ImagesStatus, in.ImagesStatus)
+			log.Println(msg)
+			return nil, NewError(errors.New(msg))
+		}
+	}
+
 	post, e := api.postPersister.UpdatePost(ctx, id, in)
 	if e != nil {
 		return nil, NewError(e)
 	}
 
-	if topic != nil && msg != nil {
+	if recordsWriterTopic != nil && msg != nil {
 		// send the message
-		err = topic.Send(ctx, &pubsub.Message{Body: msg})
+		err = recordsWriterTopic.Send(ctx, &pubsub.Message{Body: msg})
 		if err != nil { // this had best never happen
 			log.Printf("[ERROR] Can't send message %v", err)
 			// undo the update
@@ -180,34 +287,63 @@ func (api API) UpdatePost(ctx context.Context, id uint32, in model.Post) (*model
 	}
 
 	if currPost.RecordsKey != "" && currPost.RecordsKey != in.RecordsKey {
-		api.deleteRecordsData(ctx, currPost.RecordsKey)
+		api.deleteReferencedContent(ctx, currPost.RecordsKey)
 	}
+
+	if imagesWriterTopic != nil && msg != nil {
+		// send the message
+		err = imagesWriterTopic.Send(ctx, &pubsub.Message{Body: msg})
+		if err != nil { // this had best never happen
+			log.Printf("[ERROR] Can't send message %v", err)
+			// undo the update
+			_, _ = api.postPersister.UpdatePost(ctx, id, *currPost)
+			return nil, NewError(err)
+		}
+	}
+
+	if currPost.ImagesKey != "" && currPost.ImagesKey != in.ImagesKey {
+		api.deleteReferencedContent(ctx, currPost.ImagesKey+".zip")
+	}
+
 	return post, nil
 }
 
 // DeletePost holds the business logic around deleting a Post
 func (api API) DeletePost(ctx context.Context, id uint32) error {
+	log.Printf("[DEBUG] deleting %d", id)
 	post, err := api.GetPost(ctx, id)
 	if err != nil {
 		return NewError(err)
 	}
 	if post.RecordsStatus != model.PostDraft {
-		return NewError(fmt.Errorf("post must be in Draft status; is %s", post.RecordsStatus))
+		return NewError(fmt.Errorf("post(%d).RecordsStatus must be Draft; is %s", post.ID, post.RecordsStatus))
 	}
+	if post.ImagesStatus != model.PostDraft && post.ImagesStatus != "" /* backwards compatibility */ {
+		return NewError(fmt.Errorf("post(%d).ImagesStatus must be Draft; is %s", post.ID, post.ImagesStatus))
+	}
+	log.Printf("[DEBUG] deleting records for %d", id)
 	// delete records for post first so we don't have referential integrity errors
 	if err := api.DeleteRecordsForPost(ctx, id); err != nil {
 		return err
 	}
+	if post.ImagesKey != "" {
+		log.Printf("[DEBUG] deleting images for %d", id)
+		if err := api.DeleteImagesForPost(ctx, id); err != nil {
+			return NewError(err)
+		}
+	}
+	log.Printf("[DEBUG] deleting post %d", id)
 	if err := api.postPersister.DeletePost(ctx, id); err != nil {
 		return NewError(err)
 	}
+	log.Printf("[DEBUG] deleting content for %d", id)
 	if post.RecordsKey != "" {
-		api.deleteRecordsData(ctx, post.RecordsKey)
+		api.deleteReferencedContent(ctx, post.RecordsKey)
 	}
 	return nil
 }
 
-func (api API) deleteRecordsData(ctx context.Context, key string) {
+func (api API) deleteReferencedContent(ctx context.Context, key string) {
 	// delete records data
 	bucket, err := api.OpenBucket(ctx)
 	if err != nil {
@@ -217,4 +353,36 @@ func (api API) deleteRecordsData(ctx context.Context, key string) {
 	if err := bucket.Delete(ctx, key); err != nil {
 		log.Printf("[ERROR] error deleting records file %v\n", err)
 	}
+}
+
+// DeleteImagesForPost holds the business logic around deleting the images for a Post
+func (api API) DeleteImagesForPost(ctx context.Context, postID uint32) error {
+	post, err := api.GetPost(ctx, postID)
+	if err != nil {
+		return NewError(err)
+	}
+	bucket, err := api.OpenBucket(ctx)
+	if err != nil {
+		log.Printf("[ERROR] OpenBucket %v\n", err)
+		return NewError(err)
+	}
+	defer bucket.Close()
+	li := bucket.List(&blob.ListOptions{
+		Prefix: post.ImagesKey,
+	})
+	for {
+		obj, err := li.Next(ctx)
+		if err == io.EOF {
+			break
+		} else if err != nil {
+			return NewError(err)
+		}
+		log.Printf("[DEBUG] Deleting key %s", obj.Key)
+		err = bucket.Delete(ctx, obj.Key)
+		if err != nil {
+			log.Printf("[ERROR] Error deleting key %s", obj.Key)
+			return NewError(err)
+		}
+	}
+	return nil
 }
